pyproject.toml
[tool.poetry]
name = "FeatureFuse"
version = "0.1.0"
description = "make you generate specific features based on config file"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
license = "MIT"
packages = [
  { include = "featurefuse" }
]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.3"

[tool.poetry.group.dev.dependencies]
hydra-core = "^1.1.1"
pyyaml = "^6.0"
pre-commit = "^3.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
